honor MONGOOSE_SKIP_DROP on all db drops, allows testing on existing test db with user credentials (caveats: leaves a mess, and multi_client test probably won't work because requires permission to create db)

+++ t/zdone.t
@@ -6,7 +6,8 @@ use lib 't/lib';
 use MongooseT; # this connects to the db for me
 my $db = db;
 
-my $ret = $db->run_command({  'dropDatabase' => 1  }); 
-ok( 1, 'dropped' );   # dont really care if it's dropped
+
+my $ret = $db->run_command({  'dropDatabase' => 1  }) unless $ENV{MONGOOSE_SKIP_DROP}; 
+ok( 1, 'done' );   # dont really care if it's dropped
 
 done_testing;
